refactor(media): extract usePosterCascade to shared hook, reuse in MediaCard and MovieHero
Moves the 3-tier poster fallback logic (posterUrl → fallbackPosterUrl → placeholder)
from an inline function in MediaCard into a shared hooks/usePosterCascade.ts, and
updates MovieHero's HeroPoster to use the same hook (2-tier path). Adds a useEffect
reset so state clears whenever the primary URL changes (preserves existing MovieHero
test coverage for prop-change behaviour). Closes #2404.
